Modul Pertemuan 5: Otorisasi (Authorization) - Role, Gate, dan Policy

+++ app/Http/Controllers/ProductController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Product;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ProductController extends Controller
 {
@@ -47,6 +48,9 @@ class ProductController extends Controller
     {
         $product = Product::findOrFail($id);
 
+        //Cek apakah user berhak mengupdate produk ini
+        Gate::authorize('update', $product);
+
         $validated = $request->validate([
             'name' => 'sometimes|string|max:255',
             'quantity' => 'sometimes|integer',
@@ -61,6 +65,9 @@ class ProductController extends Controller
 
     public function edit(Product $product)
     {
+        //Cek apakah user berhak melihat halaman edit produk ini
+        Gate::authorize('update', $product);
+
         $users = User::orderBy('name')->get();
 
         return view('product.edit', compact('product', 'users'));
@@ -70,8 +77,16 @@ class ProductController extends Controller
     {
         $product = Product::findOrFail($id);
 
+        //Cek apakah user berhak menghapus produk ini
+        Gate::authorize('delete', $product);
+
         $product->delete();
 
         return redirect()->route('product.index')->with('success', 'Product berhasil dihapus');
+    }
+
+    public function export()
+    {
+        return "Berhasil!";
     }
 }
